feat: separate benchmark plotting and align MOA options
